moved playback_rates const into separate file
