+++ packages/components/media-player/src/PLAYBACK_RATES.js
@@ -1,0 +1,16 @@
+const PLAYBACK_RATES = [
+  { value: 0.2, label: '0.2' },
+  { value: 0.25, label: '0.25' },
+  { value: 0.5, label: '0.5' },
+  { value: 0.75, label: '0.75' },
+  { value: 1, label: '1' },
+  { value: 1.25, label: '1.25' },
+  { value: 1.5, label: '1.5' },
+  { value: 1.75, label: '1.75' },
+  { value: 2, label: '2' },
+  { value: 2.5, label: '2.5' },
+  { value: 3, label: '3' },
+  { value: 3.5, label: '3.5' }
+];
+
+export default PLAYBACK_RATES;
